Add test application to check the current natvis
This test application is creating multiple types that are (or are not
yet) handled by the natvis file.

How to use:
- open the project with Visual Studio
- put a break point on the return, in the main.cpp
- run in debug

The CMakeFiles.txt is at the root, so when opening the project with
Visual Studio, one can edit the qt6.natvis file directly to see changes.
In order for that to work, we need to remove the ##NAMESPACE##:: from
the natvis file.

Note: only tested with Qt6, I'm not sure it's worth having a test app
for Qt5.

Change-Id: I2ab544d91dfc88a16784189ce5c789987b1f1886
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>